refactor(io): separate inputs and outputs into two submodules
