+++ queenbee/io/inputs/__init__.py
@@ -1,0 +1,1 @@
+"""Input objects."""
