+++ queenbee/io/outputs/__init__.py
@@ -1,0 +1,1 @@
+"""Output objects."""
