+++ queenbee/plugin/function.py
@@ -3,7 +3,8 @@ from typing import List
 from pydantic import Field, validator, constr
 
 from ..io.common import IOBase
-from ..io.function import FunctionInputs, FunctionOutputs
+from ..io.inputs.function import FunctionInputs
+from ..io.outputs.function import FunctionOutputs
 from ..base.variable import validate_inputs_outputs_var_format, get_ref_variable
